+++ src/timelines/vertical.ts
@@ -1,5 +1,11 @@
+import { sortAndRenderNestedEvents } from '.'
 import { AllNotesData, CardContainer, DivWithCalcFunc } from '../types'
-import { buildTimelineDate, createInternalLinkOnNoteCard, handleColor } from '../utils'
+import {
+  buildMinimizedDateString,
+  buildTimelineDate,
+  createInternalLinkOnNoteCard,
+  handleColor,
+} from '../utils'
 
 /**
    * Build a vertical timeline
@@ -18,106 +24,143 @@ export async function buildVerticalTimeline(
   let eventCount = 0
   // Build the timeline html element
   for ( const date of timelineDates ) {
-    const align   = eventCount % 2 === 0 ? 'left' : 'right'
-    const datedTo = [ timelineNotes[date][0].startDate.replace( /-0*$/g, '' )
-                  + ( timelineNotes[date][0].era ? ` ${timelineNotes[date][0].era}` : '' )
-    ]
+    if ( !timelineNotes[date] ) {
+      continue
+    }
 
-    const noteDivs = [
-      timeline.createDiv(
-        { cls: ['timeline-container', `timeline-${align}`] }, 
-        ( div ) => {
-          div.style.setProperty( '--timeline-indent', '0' )
-          div.setAttribute( 'timeline-date', timelineNotes[date][0].startDate )
-          div.setAttribute( 'collapsed', 'false' )
-        }
-      )
-    ]
+    const align = eventCount % 2 === 0 ? 'left' : 'right'
 
-    const eventsDiv = noteDivs[0].createDiv({
+    const firstNote = timelineNotes[date][0]
+    const firstDate = firstNote.startDate
+    const {
+      readable: readableStartDate,
+      normalized: normalizedStartDate,
+    } = buildMinimizedDateString( firstDate )
+    const datedTo = [ readableStartDate + ( firstDate.era ? ` ${firstDate.era}` : '' )]
+
+    const noteDivs: HTMLDivElement[] = []
+    const containerDiv = timeline.createDiv(
+      { cls: ['timeline-container', `timeline-${align}`] },
+      ( div ) => {
+        div.style.setProperty( '--timeline-indent', '0' )
+        div.setAttribute( 'timeline-date', normalizedStartDate )
+        div.setAttribute( 'collapsed', 'false' )
+      }
+    )
+
+    noteDivs.push( containerDiv )
+
+    const eventsDiv = containerDiv.createDiv({
       cls: 'timeline-event-list',
-      attr: { 'style': 'display: block' }
+      attr: { style: 'display: block' }
     })
 
-    const noteHdrs = [noteDivs[0].createEl( 'h2', {
-      attr: { 'style' : `text-align: ${align};` },
-      text: datedTo[0]
-    })]
+    containerDiv.createEl( 'h2', {
+      attr: { style: `text-align: ${align};` },
+      text: readableStartDate
+    })
 
     for ( const rawNote of timelineNotes[date] ) {
       // for confirmation of types
       const note: CardContainer = rawNote
-      const { endDate: endDateString, era } = note
-      const startDate = buildTimelineDate( note.startDate )
-      const endDate = buildTimelineDate( endDateString )
+      const { endDate: end, era, startDate: start, type } = note
+      const startDate: Date | null | string = buildTimelineDate( start )
+      let endDate: Date | null | string     = buildTimelineDate( end )
 
-      // skip events that are of type 'box' or 'point', or if the endDate is invalid
-      if ( ['box', 'point'].includes( note.type ) || endDate < startDate ) {
+      if ( !startDate ) {
+        console.error( "Couldn't build a start date for timeline on vertical timeline" )
         continue
       }
 
-      if ( !noteDivs[0].classList.contains( 'timeline-head' )) {
-        noteDivs[0].classList.add( 'timeline-head' ) 
+      if ( !endDate ) {
+        endDate = ''
       }
 
-      const dated = endDateString + ( era ?? '' )
-      const lastTimelineDate = buildTimelineDate( noteDivs.last().getAttribute( 'timeline-date' ))
+      // skip events that are of type 'box' or 'point', or if the endDate is invalid
+      if ( ['box', 'point'].includes( type ) || endDate < startDate  || !noteDivs ) {
+        continue
+      }
+
+      if ( !containerDiv.classList.contains( 'timeline-head' )) {
+        containerDiv.classList.add( 'timeline-head' )
+      }
+
+      const {
+        readable: readableEndDate,
+        normalized: normalizedEndDate
+      } = buildMinimizedDateString( end )
+      const endDateFormatted = readableEndDate + ( era ? ` ${era}` : '' )
+
+
+      const lastTimelineDate = buildTimelineDate( noteDivs[noteDivs.length - 1].getAttribute( 'timeline-date' ))
+      if ( !lastTimelineDate ) {
+        console.error( "There's no last timeline date" )
+        continue // change to break?
+      }
+
       if ( !datedTo[1] || endDate > lastTimelineDate ) {
-        datedTo[1] = datedTo[0] + ' to ' + dated
+        datedTo[1] = `${datedTo[0]} to ${endDateFormatted}`
       }
 
       const noteDiv = timeline.createDiv(
         { cls: ['timeline-container', `timeline-${align}`, 'timeline-tail'] },
         ( div ) => {
           div.style.setProperty( '--timeline-indent', '0' )
-          div.setAttribute( 'timeline-date', endDateString )
+          div.setAttribute( 'timeline-date', normalizedEndDate )
         }
       )
-      
-      const noteHdr = noteDiv.createEl( 'h2', {
-        attr: { 'style' : `text-align: ${align};` },
-        text: dated
+
+      noteDiv.createEl( 'h2', {
+        attr: { style: `text-align: ${align};` },
+        text: endDateFormatted
       })
-
-      noteHdrs.push( noteHdr )
       noteDivs.push( noteDiv )
-      for ( const n of noteDivs ) {
-        const timelineDate = buildTimelineDate( n.getAttribute( 'timeline-date' ))
 
-        if ( timelineDate > endDate ) {
-          n.before( noteDiv )
+      for ( let i = noteDivs.length - 2; i >= 0; i-- ) {
+        const timelineDate = buildTimelineDate( noteDivs[i].getAttribute( 'timeline-date' ))
+        if ( timelineDate && ( timelineDate > endDate )) {
+          noteDivs[i].before( noteDiv )
         }
       }
 
       /* The CSS 'timeline-timespan-length' determines both the length of the event timeline widget on the timeline,
            and the length of the vertical segment that spans between the displayed dates. If this value is not recalculated
            then these elements will not be responsive to layout changes. */
-      ( noteDivs.last() as DivWithCalcFunc ).calcLength = () => {
-        const axisMin = noteDivs[0].getBoundingClientRect().top
-        const axisMax = noteDiv    .getBoundingClientRect().top
-        const spanMin = noteHdrs[0].getBoundingClientRect().top
-        const spanMax = noteHdr    .getBoundingClientRect().top
+      ( noteDiv as DivWithCalcFunc ).calcLength = () => {
+        const axisMin = containerDiv.getBoundingClientRect().top
+        const axisMax = noteDiv.getBoundingClientRect().top
+        const spanMin = containerDiv.querySelector( 'h2' )?.getBoundingClientRect().top ?? 0
+        const spanMax = noteDiv.querySelector( 'h2' )?.getBoundingClientRect().top ?? 0
 
-        noteDivs[0].style.setProperty( '--timeline-span-length', `${axisMax - axisMin}px` )
-        noteDiv    .style.setProperty( '--timeline-span-length', `${spanMax - spanMin}px` )
+        containerDiv.style.setProperty( '--timeline-span-length', `${axisMax - axisMin}px` )
+        noteDiv     .style.setProperty( '--timeline-span-length', `${spanMax - spanMin}px` )
       }
     }
 
-    noteDivs[0].addEventListener( 'click', ( event ) => {
+    containerDiv.addEventListener( 'click', ( event ) => {
       event.preventDefault()
 
-      const collapsed = !JSON.parse( noteDivs[0].getAttribute( 'collapsed' ))
-      noteDivs[0].setAttribute( 'collapsed', String( collapsed ))
-      for ( const p of noteDivs[0].getElementsByTagName( 'p' )) {
-        p.setCssProps({ 'display': collapsed ? 'none' : 'block' })
-      }
+      const collapsed = !JSON.parse( containerDiv.getAttribute( 'collapsed' ) ?? '{}' )
+      containerDiv.setAttribute( 'collapsed', String( collapsed ))
+
+      Array.from( containerDiv.getElementsByTagName( 'p' )).forEach(( p ) => {
+        p.style.display = collapsed ? 'none' : 'block'
+      })
+
       /* If this event has a duration (and thus has an end note), we hide all elements between the start and end
          note along with the end note itself */
-      if ( noteDivs.length > 0 ) {
-        noteHdrs[0].setText( collapsed && datedTo[1] ? datedTo[1] : datedTo[0] )
-        const notes = [...timeline.children]
-        const inner = notes.slice( notes.indexOf( noteDivs.first()) + 1, notes.indexOf( noteDivs.last()) + 1 )
-        inner.forEach(( note: DivWithCalcFunc ) => {
+      if ( noteDivs.length > 1 ) {
+        const h2 = containerDiv.querySelector( 'h2' )
+        if ( !h2 ) {
+          return
+        }
+
+        h2.innerText = collapsed && datedTo[1] ? datedTo[1] : datedTo[0]
+        const innerNotes = Array.from( timeline.children ).slice(
+          Array.from( timeline.children ).indexOf( noteDivs[0] ) + 1,
+          Array.from( timeline.children ).indexOf( noteDivs[noteDivs.length - 1] ) + 1
+        )
+        innerNotes.forEach(( note: DivWithCalcFunc ) => {
           note.style.display = collapsed ? 'none' : 'block'
           note.calcLength?.()
         })
@@ -126,12 +169,14 @@ export async function buildVerticalTimeline(
       /* The CSS '--timeline-indent' variable allows for scaling down the event when contained within another event,
          but in this case it also tells us how many time spanning events have had their length altered as a consequence
          of this note's mutation. */
-      let nested = +noteDivs[0].style.getPropertyValue( '--timeline-indent' ) + noteDivs.length - 1
-      for ( let f = 'nextElementSibling', sibling = noteDivs[0][f]; nested > 0; sibling = sibling[f] ) {
+      let nested = +containerDiv.style.getPropertyValue( '--timeline-indent' ) + noteDivs.length - 1
+      let sibling = containerDiv.nextElementSibling
+      while ( nested > 0 && sibling ) {
         if ( sibling.classList.contains( 'timeline-tail' )) {
-          sibling.calcLength?.()
+          ( sibling as DivWithCalcFunc ).calcLength?.()
           nested--
         }
+        sibling = sibling.nextElementSibling
       }
     })
 
@@ -151,14 +196,14 @@ export async function buildVerticalTimeline(
       currentStyle.setProperty( 'display', 'none' )
     })*/
 
-    if ( !timelineNotes[date] ) {
-      continue 
-    }
-
     for ( const eventAtDate of timelineNotes[date] ) {
       const noteCard = eventsDiv.createDiv({ cls: 'timeline-card' })
-      noteCard.className += ` ${eventAtDate.classes}`
-      // add an image only if available
+      createInternalLinkOnNoteCard( eventAtDate, noteCard )
+
+      if ( eventAtDate.classes ) {
+        noteCard.classList.add( eventAtDate.classes )
+      }
+
       if ( eventAtDate.img ) {
         noteCard.createDiv({
           cls: 'thumb',
@@ -166,39 +211,17 @@ export async function buildVerticalTimeline(
         })
       }
 
+      // todo : add more support for other custom classes
       if ( eventAtDate.color ) {
-        // todo : add more support for other custom classes
         handleColor( eventAtDate.color, noteCard, eventAtDate.id )
       }
 
-      createInternalLinkOnNoteCard( eventAtDate, noteCard )
-      eventAtDate.body && noteCard.createEl( 'p', { text: eventAtDate.body.trim() })
-    }
-
-    /* If the previous note is of class 'timeline-tail' then there's a chance that the current node belongs nested within
-       the previous one. Here we iterate backwards for as long as elements of the class 'timeline-tail' are encountered.
-
-       Then we sort an array containing the start and ending date of the current event along with the ending date of the
-       previous event. If the index of the ending date of the previous event in the new array is greater than 0 then the
-       note(s) preceding it (either the note belonging to the start date or the notes belonging to both the start and end
-       dates of the current event) are placed before their predecessor. */
-    for ( let s = [...timeline.children],i = s.indexOf( noteDivs[0] ); s[i-1]?.classList.contains( 'timeline-tail' ); i-- ) {
-      const t = s[i-1].getAttribute( 'timeline-date' )
-      const times = [
-        t,
-        ...noteDivs.map(( n ) => {
-          return n.getAttribute( 'timeline-date' )
-        })
-      ]
-
-      const lastTIndex = times.sort().lastIndexOf( t )
-      for ( let j = lastTIndex; j > 0; s[i-1].before( ...noteDivs.slice( 0, j )), j = 0 ) {
-        const indent = +noteDivs[0].style.getPropertyValue( '--timeline-indent' ) + 1
-        noteDivs.forEach(( n ) => {
-          n.style.setProperty( '--timeline-indent', `${ indent }` )
-        })
+      if ( eventAtDate.body ) {
+        noteCard.createEl( 'p', { text: eventAtDate.body.trim() })
       }
     }
+
+    sortAndRenderNestedEvents( noteDivs, timeline )
 
     eventCount++
   }