+++ src/types.ts
@@ -1,4 +1,5 @@
 import { FrontMatterCache, MetadataCache, TFile, Vault } from 'obsidian'
+import { DataItem } from 'vis-timeline'
 
 /* ------------------------------ */
 /*              Enums             */
@@ -54,7 +55,7 @@ export interface EventItem {
   id: number,
   className: string,
   content: string,
-  end: Date,
+  end: Date | undefined,
   path: string,
   start: Date,
   type: string,
@@ -100,6 +101,16 @@ export interface InternalTimelineArgs {
   zoomOutLimit: number,
 }
 
+export interface MinimizedResult {
+  readable: string,
+  cleaned: string,
+  normalized: string,
+}
+
+export interface NormalizeAndCleanDateOutput extends CleanedDateResultObject {
+  normalizedDate: string
+}
+
 export interface TimelinesSettings {
   eventElement: AcceptableEventElements,
   frontMatterKeys: FrontMatterKeys,
@@ -115,5 +126,6 @@ export interface TimelinesSettings {
 /* ------------------------------ */
 
 export type AllNotesData = ( CardContainer[] )[]
+export type CombinedTimelineEventData = EventItem & DataItem
 export type DivWithCalcFunc = HTMLDivElement & { calcLength?: () => void }
 export type EventCountData = ( HTMLElement | FrontMatterCache | null )[]