+++ src/utils/dates.ts
@@ -2,7 +2,73 @@ import { DateTime } from 'luxon'
 
 import { logger } from './debug'
 import { DEFAULT_SETTINGS } from '../constants'
-import { CleanedDateResultObject } from '../types'
+import { CleanedDateResultObject, MinimizedResult, NormalizeAndCleanDateOutput } from '../types'
+
+export const buildMinimizedDateString = ( str: string ): MinimizedResult => {
+  const normalizedAndCleaned = normalizeAndCleanDate( str )
+  if ( !normalizedAndCleaned ) {
+    throw new Error( `Could not normalize and/or clean the date: ${str}` )
+  }
+
+  const { cleanedDateString: cleaned, normalizedDate: normalized } = normalizedAndCleaned
+  const readable = minimizeDateString( cleaned )
+
+  const result: MinimizedResult = {
+    readable,
+    cleaned,
+    normalized,
+  }
+
+  return result
+}
+
+/** 
+ * @param dateString
+ * 
+ * @returns {string}
+ */
+const minimizeDateString = ( dateString: string ): string => {
+  const chars = Array.from( dateString )
+
+  let isNegative = false
+  if ( chars[0] === '-' ) {
+    isNegative = true
+  }
+
+  let sections: string[] = []
+  if ( isNegative ) {
+    sections = chars.slice( 1, chars.length ).join( '' ).split( '-' )
+  } else {
+    sections = chars.join( '' ).split( '-' )
+  }
+
+  if ( !sections.length ) {
+    throw new Error( 'could not get the different sections of the event date' )
+  }
+
+  logger( 'minimizeDateString | before', { sections })
+
+  // we always at least have the year
+  const remainingSections: string[] = [ (( isNegative ? '-' : '' ) + sections[0] ) ]
+  for ( let i = 1; i < 4; i++ ) {
+    // if whatever section we're looking at is invalid, skip
+    if ( [0, -1].includes( parseInt( sections[i] ))) {
+      continue
+    }
+
+    // if we're looking at hour but day is not set, skip
+    // todo: if hour is not set, but day is, don't show it
+    if ( i === 3 && parseInt( sections[i - 1] ) < 1 ) {
+      continue
+    }
+
+    remainingSections.push( sections[i] )
+  }
+
+  logger( 'minimizeDateString | remaining', { remainingSections })
+
+  return remainingSections.join( '-' )
+}
 
 /**
  * Take a normalizedDate and clean it of leading zeros, the return all the various
@@ -12,7 +78,10 @@ import { CleanedDateResultObject } from '../types'
  * 
  * @returns {CleanedDateResultObject}
  */
-export const cleanDate = ( normalizedDate: string ): CleanedDateResultObject => {
+export const cleanDate = ( normalizedDate: string ): CleanedDateResultObject | null => {
+  if ( normalizedDate === null ) {
+    return null
+  }
   const isNegative = normalizedDate[0] === '-'
   const parts = normalizedDate.slice( 1 ).split( '-' )
 
@@ -22,7 +91,7 @@ export const cleanDate = ( normalizedDate: string ): CleanedDateResultObject => 
   
   const cleanedDateString = isNegative ? '-' + numParts.join( '-' ) : numParts.join( '-' )
   const year  = numParts[0] * ( isNegative ? -1 : 1 )
-  const month = ( numParts[1] ?? 1 ) - 1
+  const month = ( numParts[1] ?? 1 ) - ( numParts[0] !== 0 ? 1 : 0 )
   const day   = numParts[2]
   const hour  = numParts[3] ?? 1
 
@@ -37,6 +106,26 @@ export const cleanDate = ( normalizedDate: string ): CleanedDateResultObject => 
   return resultObject
 }
 
+export const normalizeAndCleanDate = (
+  date: string | null,
+  maxDigits: number = parseInt( DEFAULT_SETTINGS.maxDigits )
+): NormalizeAndCleanDateOutput | null => {
+  const normalizedDate = normalizeDate( date, maxDigits )
+  if ( !normalizedDate ) {
+    return null
+  }
+  
+  const cleanedDateObject = cleanDate( normalizedDate )
+  if ( !cleanedDateObject ) {
+    return null
+  }
+
+  return {
+    ...cleanedDateObject,
+    normalizedDate
+  }
+}
+
 /**
  * Takes a date string and normalizes it so there are always 4 sections, each the length specified by maxDigits
  * If there are missing sections, they will be inserted with a value of 01 (except for hours, which will be 00)
@@ -47,7 +136,7 @@ export const cleanDate = ( normalizedDate: string ): CleanedDateResultObject => 
  * @returns {string}
  */
 export const normalizeDate = (
-  date: string,
+  date: string | null,
   maxDigits: number = parseInt( DEFAULT_SETTINGS.maxDigits )
 ): string | null => {
   if ( !date ) {
@@ -97,26 +186,21 @@ export const normalizeDate = (
  * @returns {Date | null}
  */
 export const buildTimelineDate = (
-  rawDate: string,
+  rawDate: string | null,
   maxDigits?: number
 ): Date | null => {
-  const normalizedDate = normalizeDate( rawDate, maxDigits ?? parseInt( DEFAULT_SETTINGS.maxDigits ))
-  if ( !normalizedDate ) {
+  const normalizedAndCleanedDateObject = normalizeAndCleanDate( rawDate, maxDigits )
+  if ( !normalizedAndCleanedDateObject ) {
     return null
   }
-  
-  const cleanedDateObject = cleanDate( normalizedDate )
-  if ( !normalizedDate ) {
-    return null
-  }
-  const { cleanedDateString, year, month, day, hour } = cleanedDateObject
+  const { cleanedDateString, year, month, day, hour, normalizedDate } = normalizedAndCleanedDateObject
 
   // native JS Date handles negative years and recent dates pretty decent
   // so if year is negative, or if the year is recent (past 1900)
   // we can just use the JS Date directly with no workarounds
   let returnDate: Date
-  let luxonDateTime: DateTime
-  let luxonDateString: string
+  let luxonDateTime: DateTime | null = null
+  let luxonDateString: string | null = null
   if ( year < 0 || year > 1900 ) {
     returnDate = new Date( year, month, day, hour )
   } else {
@@ -124,13 +208,20 @@ export const buildTimelineDate = (
     // the date that we want. JS Date will change "0001-00-01" to "Jan 1st 1970"
     luxonDateTime = DateTime.fromFormat( cleanedDateString, 'y-M-d-H' )
     luxonDateString = luxonDateTime.toISO()
+
+    if ( !luxonDateString ) {
+      console.error( "Couldn't create a luxon date string!" )
+      return null
+    }
+
     returnDate = new Date( luxonDateString )
   }
 
   logger( 'buildTimelineDate | date variables', {
     rawDate,
-    normalizedDate,
     cleanedDateString,
+    normalizedAndCleanedDateObject,
+    normalizedDate,
     luxonDateTime,
     luxonDateString,
     returnDate