+++ src/utils/events.ts
@@ -1,7 +1,10 @@
 import { FrontMatterCache, MetadataCache, Notice, TFile, Vault, normalizePath } from 'obsidian'
+import { DataItem, IdType } from 'vis-timeline'
+
 import {
   EventCountData,
   EventDataObject,
+  EventItem,
   EventTypeNumbers,
   FrontMatterKeys,
   GetFileDataInput
@@ -16,6 +19,55 @@ export const isFrontMatterCacheType = ( value: unknown ): value is FrontMatterCa
 
 export const isHTMLElementType = ( value: unknown ): value is HTMLElement => {
   return value instanceof HTMLElement
+}
+
+export const buildCombinedTimelineDataObject = ( event: EventItem, obj: object = {}) => {
+  return {
+    ...buildBaseDataItem(),
+    ...event,
+    ...obj
+  }
+}
+
+const buildBaseDataItem = (): Omit<DataItem, 'id'> & { id: IdType } => {
+  // export interface DataItem {
+  //   className?: string;
+  //   content: string;
+  //   end?: Date | number | string;
+  //   group?: any;
+  //   id?: string | number;
+  //   start: Date | number | string;
+  //   style?: string;
+  //   subgroup?: string | number;
+  //   title?: string;
+  //   type?: string;
+  //   editable?: boolean | {
+  //     remove?: boolean;
+  //     updateGroup?: boolean;
+  //     updateTime?: boolean;
+  //   }
+  //   selectable?: boolean;
+  //   limitSize?: boolean;
+  // }
+
+  const baseDataItem: Omit<DataItem, 'id'> & { id: IdType } = {
+    // skipped optional keys that will be provided by the event 
+    // className: eventItem.className,
+    // end: eventItem.end ?? '',
+
+    content: '', // will be overwritten by the event content
+    id:      '', // will be overwritten by the event id
+    start:   '', // will be overwritten by the event start
+
+    editable:  false,
+    group: undefined,
+    limitSize: false,
+    selectable: true,
+    style: undefined,
+    title: undefined,
+  }
+
+  return baseDataItem
 }
 
 /**
@@ -39,22 +91,22 @@ export async function getNumEventsInFile(
   }
 
   // even though there should only ever be 1, we still filter so that we get back an array
-  const frontMatter = combinedEventsAndFrontMatter.filter( isFrontMatterCacheType )
-  const events = combinedEventsAndFrontMatter.filter( isHTMLElementType )
+  const frontMatter = combinedEventsAndFrontMatter?.filter( isFrontMatterCacheType )
+  const events = combinedEventsAndFrontMatter?.filter( isHTMLElementType )
 
   logger( 'getNumEventsInFile | events & frontmatter', { events, frontMatter })
-  const numFrontMatter = frontMatter.length
-  const numEvents = events.length
+  const numFrontMatter = frontMatter?.length ?? 0
+  const numEvents = events?.length ?? 0
 
   return { numEvents, numFrontMatter, totalEvents: numEvents + numFrontMatter }
 }
 
 export const getEventsInFile = async (
-  file: TFile,
-  appVault: Vault,
-  fileCache: MetadataCache
+  file: TFile | null | undefined,
+  appVault: Vault | null | undefined,
+  fileCache: MetadataCache | null | undefined
 ): Promise<EventCountData | null> => {
-  if ( !file ) {
+  if ( !file || !appVault || !fileCache ) {
     return null
   }
 
@@ -63,7 +115,7 @@ export const getEventsInFile = async (
   const rawEvents = doc.getElementsByClassName( 'ob-timelines' )
   fileEvents.push( ...Array.from( rawEvents ).filter( isHTMLElementType ))
 
-  const frontMatterData = fileCache.getFileCache( file ).frontmatter
+  const frontMatterData = fileCache.getFileCache( file )?.frontmatter
   if ( frontMatterData ) {
     fileEvents.push( frontMatterData )
   }
@@ -80,7 +132,7 @@ export const getEventData = (
 ): EventDataObject | null => {
   logger( 'getEventData | function starting for eventObject:', eventObject )
   const startDate = retrieveEventValue(
-    eventObject, 'startDate', null, frontMatterKeys?.startDateKey
+    eventObject, 'startDate', '', frontMatterKeys?.startDateKey
   )
   if ( !startDate ) {
     new Notice( `No date found for ${file.name}` )
@@ -93,16 +145,16 @@ export const getEventData = (
   const endDate        = retrieveEventValue(
     eventObject, 'endDate', startDate, frontMatterKeys?.endDateKey
   )
-  const era            = retrieveEventValue( eventObject, 'era', null )
-  const eventImg       = retrieveEventValue( eventObject, 'img', null )
+  const era            = retrieveEventValue( eventObject, 'era', '' )
+  const eventImg       = retrieveEventValue( eventObject, 'img', '' )
   const noteBody       = retrieveEventValue( eventObject, 'description', defaultBody )
   const notePath       = retrieveEventValue( eventObject, 'path', '/' + normalizePath( file.path ))
   const noteTitle      = retrieveEventValue(
     eventObject, 'title', file.name.replace( '.md', '' ), frontMatterKeys?.titleKey
   )
-  const tags           = retrieveEventValue( eventObject, 'tags', '' )
+  const tags           = retrieveEventValue( eventObject, 'tags', '' ) ?? ''
   const type           = retrieveEventValue( eventObject, 'type', 'box' )
-  const showOnTimeline = retrieveEventValue( eventObject, 'showOnTimeline', null )
+  const showOnTimeline = retrieveEventValue( eventObject, 'showOnTimeline', 'false' )
 
   const eventData: EventDataObject = {
     classes,
@@ -125,9 +177,9 @@ export const getEventData = (
 const retrieveEventValue = (
   eventData: HTMLElement | FrontMatterCache,
   datasetKey: string,
-  defaultValue: string | null,
+  defaultValue: string,
   frontMatterKeys?: string[] | null,
-): string | null => {
+): string => {
   if ( isHTMLElementType( eventData )) {
     return retrieveHTMLValue( eventData, datasetKey, defaultValue )
   } else {
@@ -138,23 +190,28 @@ const retrieveEventValue = (
 const retrieveHTMLValue = (
   event: HTMLElement,
   datasetKey: string,
-  defaultValue: string | null,
-): string | null => {
-  logger( 'retrieveHTMLValue | datasetKey:', datasetKey )
+  defaultValue: string = '',
+): string => {
+  logger( 'retrieveHTMLValue | datasetKey:', { key: datasetKey, value: event.dataset[datasetKey], defaultValue })
   const result = event.dataset[datasetKey]
 
-  return result ?? defaultValue
+  if ( !result || result === '' ) {
+    return defaultValue
+  }
+
+  return result
 }
 
 const retrieveFrontMatterValue = (
   event: FrontMatterCache,
   datasetKey: string,
-  defaultValue: string | null,
+  defaultValue: string = '',
   frontMatterKeys?: string[] | null,
-): string | null => {
+): string => {
   logger( 'retrieveFrontMatterValue | datasetKey:', datasetKey )
+  const alternativeValue = frontMatterKeys && findMatchingFrontMatterKey( event, frontMatterKeys )
   const result = event[datasetKey]
-    ?? findMatchingFrontMatterKey( event, frontMatterKeys )
+    ?? alternativeValue
 
   if ( !result || result === '' ) {
     return defaultValue